Support running plugin on slave node
Original version
=================
Password is obtained on an ad-hoc manner for each build.(enhance security)
This method works well on the master but crashes on the slave.
Reason is the obtained raw password is going thru encryption phase
and saved in persistence storage.
However since the slave doesn’t have the credential store it crashes

Fork version
============
Password is obtained during the jenkins master init phase.
Password is saved to the credential store for subsequence builds.
The password is used by the slaves nodes in each build.

Obtaining the password during each build is don